Added TStr to unsigned int conversion.
Added the ToUsize trait to convert TStr to u128 and usize.

Added docsrs feature (forgot to do it in a previous commit).

Added `[package.metadata.docs.rs]` to the Cargo.toml, automatically enable features in docs.rs

Changed the tuple-based representation of TStr to use singly-nested tuples of tuples
(the inner ones are up to 8 elements long).