Made syn/proc_macro2 opt-in dependencies.
Added "syn_" and "proc_macro2_" features to use those dependencies.

In order to make syn an opt-in dependency, I added parsing of integer and string literals,
that's used when syn is disabled.

Changed imports to use items from proc_macro or proc_macro2 depending on the "proc_macro2_" feature.

Created the dependency-free parser in the non_syn_parsing module.

Moved the syn dependent parsing to the use_syn module.